Finish README.md
As I have said that racism is funny. But htis does not in any way immplies that I am a racist person. My observation is simply based on the fact that I have studied a few many jokes and all of them have some specific bias towards a specific type of  people which derives its root in racism